+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Convert HA to use individual inputs messages (#253)
 * New configuration open publish_inputs_all_trigger (#256)
 * Move time register messages into new parser (start+end times for AC Charge etc) (#259)
+* Publish start-end time message when time registers change externally (#261)
 
 
 # 0.13.0 - 27th October 2023